chore: reorganize unit tests in container-registry-v5
