+++ packages/container-registry-v5/package.json
@@ -57,7 +57,7 @@
     "postpublish": "rm oclif.manifest.json",
     "posttest": "yarn lint",
     "prepack": "oclif manifest",
-    "test": "cross-env TZ=utc nyc mocha",
+    "test": "cross-env TZ=utc nyc mocha --forbid-only \"test/**/*.unit.test.js\"",
     "version": "oclif readme && git add README.md"
   },
   "topic": "container"
