+++ packages/container-registry-v5/test/unit/commands/release.unit.test.js
@@ -2,7 +2,7 @@
 /* globals beforeEach afterEach */
 
 const cli = require('heroku-cli-util')
-const cmd = require('../..').commands.find(c => c.topic === 'container' && c.command === 'release')
+const cmd = require('../../..').commands.find(c => c.topic === 'container' && c.command === 'release')
 const {expect} = require('chai')
 const sinon = require('sinon')
 const nock = require('nock')
