+++ packages/container-registry-v5/test/unit/commands/run.unit.test.js
@@ -2,11 +2,11 @@
 /* globals beforeEach afterEach */
 
 const cli = require('heroku-cli-util')
-const cmd = require('../..').commands.find(c => c.topic === 'container' && c.command === 'run')
+const cmd = require('../../..').commands.find(c => c.topic === 'container' && c.command === 'run')
 const {expect} = require('chai')
 const sinon = require('sinon')
 
-const Sanbashi = require('../../lib/sanbashi')
+const Sanbashi = require('../../../lib/sanbashi')
 let sandbox
 
 describe('container run', () => {
